+++ crates/multicalc/tests/suite/linear_algebra/mod.rs
@@ -4,4 +4,5 @@ mod macros;
 mod matrix;
 mod qr;
 mod svd;
+mod symmetric_eigendecomposition;
 mod vector;